backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
python-dotenv==1.0.1
anthropic==0.40.0
openai==1.56.0
Pillow==10.4.0
